+++ R/tidywater-package.R
@@ -246,3 +246,22 @@ NULL
 #'
 #' @source Cho (2007)
 "pactoccoeffs"
+
+#' Data frame of TOC removal requirements per Stage 1 DBP Rule
+#'
+#' A dataset containing required TOC removal based on raw alkalinity and TOC
+#'
+#' @docType data
+#' @keywords datasets
+#' @name toc_compliance_table
+#' @format A dataframe with 9 rows and 5 columns
+#' \describe{
+#' \item{toc_min}{Minimum raw water TOC for compliance}
+#' \item{toc_max}{Maximum raw water TOC for compliance}
+#' \item{alk_min}{Minimum raw water alkalinity for compliance}
+#' \item{alk_max}{Maximum raw water alkalinity for compliance}
+#' \item{required_compliance}{Required percent TOC removal based on raw TOC andn alklalinity range}
+#' }
+#'
+#' @source Stage 1 Disinfection Byproduct Rule (1998)
+"toc_compliance_table"